JAVA-2260: Handle empty collections in PreparedStatement.bind(...)
Motivation:

`preparedStatement.bind(new HashSet<>())` fails with a
CodecNotFoundException.

This is because the codec registry defaults to Set<Boolean> when it
tries to infer the type of the value (this works fine for lookups by
Java type only, because all empty collections are encoded in the same
way, so any collection codec will do -- but here we happen to also know
the CQL type, and it doesn't match).

Modifications:

Special-case empty collections at the beginning of codecFor(CqlType,
value): simply look up a collection codec for an arbitrary element type.

Result:

We return a codec that can encode the provided value.
Note that it doesn't accept the CQL type, nor the declared Java type of
the value, but the contract of the method doesn't say that it should.